tests/wpt/web-platform-tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.strokeRect.globalalpha.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.strokeRect.globalalpha</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.strokeRect.globalalpha</h1>
<p class="desc">strokeRect is affected by globalAlpha</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/clear-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("strokeRect is affected by globalAlpha");
_addTest(function(canvas, ctx) {

ctx.globalAlpha = 0;
ctx.strokeStyle = '#f00';
ctx.lineWidth = 50;
ctx.strokeRect(25, 24, 50, 2);
_assertPixel(canvas, 50,25, 0,0,0,0, "50,25", "0,0,0,0");


});
</script>
